# ENH: use query_bulk in sjoin