This is the next step for #1404.

Benchmarks:
| before <master> | after <sjoin-query-bulk> | ratio | test                     | pygeos installed |
|-----------------|--------------------------|-------|--------------------------|------------------|
| 9.51±0.1s       | 178±1ms                  | ~0.02 | time_sjoin('contains')   | Y                |
| 8.40±0.4s       | 7.25±0.02s               | ~0.86 | time_sjoin('contains')   | N                |
| 9.09±0.06s      | 592±9ms                  | 0.07  | time_sjoin('intersects') | Y                |
| 7.80±0.01s      | 11.0±0.1s                | ~1.41 | time_sjoin('intersects') | N                |
| 1.14±0.02s      | 585±4ms                  | 0.51  | time_sjoin('within')     | Y                |
| 7.94±0.2s       | 11.5±0.06s               | ~1.45 | time_sjoin('within')     | N                |

We could keep a lot of the existing logic in place to avoid the slowdown without pygeos, but the way it is now reduces the complexity of the module considerably and I think will be easier to manage going forward.

If we did decide to keep the existing logic, I would like to split up `sjoin` into several functions, similar to #1271.