@jorisvandenbossche 2020-05-14
>>> This part confuses me a bit. It would be great if someone could confirm if I am right or wrong.
>>
>> As far as I understand, this would need to depend on whether the spatial index of the left or right frame is being used
>
> I don't think so. Contains is faster operation than within. And because they are interchangeable (you just change the perspective), sjoin does always contains and never within. Just flips the perspective. That is what happens in the original code above. I don't think your code does the same.

@martinfleis OK, but so with "just change the perspective", you mean: flip the left and right dataframe? (because within and contains are *not* the same, they are only the same when at the same time reversing the order of the geometries, right?)

But so, in the current sjoin code, we could switch the order of the left / right dataframe when performing the actual predicate. But now with `query_bulk`, the predicate is performed by the tree itself. So the "order" in which the geometries are tested for the predicate, depends on which tree is used? 
Which is why I would think that "flip the order of the left / right dataframe" in practice also means that you use the sindex of the left or right dataframe?