@martinfleis 2020-05-14
>  but so with "just change the perspective", you mean: flip the left and right dataframe?

Yes, that is what happens in sjoin now. `left_df, right_df = right_df, left_df`

> they are only the same when at the same time reversing the order of the geometries, right?

Correct. `polygon.contains(point)` is the same as `point.within(polygon)`. I was timing it yesterday with shapely polygons, with `contains` being about 40% faster than within. But it seems to me now, that it must have been some kind of a mistake, because I am not able to replicate those results now.

Generally, I am not much in favour of this flipping under the hood. I would prefer using directly `contains` or `within` exactly as user specified it. We may do some timings and include general guidance in docs, if one or the other way is better, but I would not do that internally.