@jorisvandenbossche 2020-05-14
> Yes, that is what happens in sjoin now. left_df, right_df = right_df, left_df

But in this PR, that line has been removed. So if we still want to keep swapping within for contains, that means we need to keep swapping the dataframes (or, swapping from which dataframe we get the spatial index).

Now, we should indeed maybe reconsider this flipping. This also relates to @brendan-ward's comment about "which sindex gets used, left or right", that it might be better to do this predictably, and then have docs for the power user explaining how to control this / what aspects can influence performance.