@adriangb 2020-05-14
For `rtree`, this swapping is still being done, but it is done in `sindex`'s `query` instead of `sjoin`. One thing to note is that `contains` supports prepared geometries but `within` does not. So if we were not doing this flipping we couldn't use prepared geometries, and I think we would see `contains` + `prep` be a lot faster than just `within`.

That said, I think it's best not to have this type of unexpected behavior. In particular, we should evaluate if there is still a performance difference under pygeos and if it is consistent across geometry types. From what I have seen, there is no consistent performance advantage. That could change in the future pygeos implements prepared geometries. If there is at some point a consistent performance advantage, then swapping should be implemented at a lower level (i.e. inside of `query` or `query_bulk`) so that all spatial index queries can take advantage of the speedup without having to swap entire dataframes and such.