@adriangb 2020-05-14
I'll also say that if we were to remove the dynamic spatial index selection, it should not break any of the tests/API. It will just make some tests/benchmarks run _very_ slow. For most users, I'm guessing that won't be a problem. For those with large datasets, it would be pretty straightforward to fix (just swap your DataFrames/parameters).