@martinfleis 2020-05-14
For a reference, there is a discussion on this topic in https://github.com/geopandas/geopandas/pull/789.