@adriangb 2020-05-14
> For a reference, there is a discussion on this topic in #789.

Thank you for the reference.

It seems that before #789, sjoin was generating it's own spatial indexes, and it would do so every time. In the discussion, you say:

> EDIT: to sum up, with this PR as it is and note in docs to generate sindex you can speed up sjoin almost 10x in some cases, 2x in other (approximately).

How much of that do you think is due to not re-creating the spatial index every time (which we of course would conserve) vs. heuristics for choosing a spatial index (which is what I think we are discussing about removing)?

For the heuristics themselves, I think there are two questions that need to be answered:

1. Is the performance delta consistent across choices of inputs and operations, for both shapely and pygeos?
2. Will this dynamic behavior become a burden if we choose to introduce more predicates later on?