@martinfleis 2020-05-14
> How much of that do you think is due to not re-creating the spatial index every time (which we of course would conserve) vs. heuristics for choosing a spatial index (which is what I think we are discussing about removing)?

Based on the benchmarks in https://github.com/geopandas/geopandas/pull/789#issuecomment-521291074 it seems to be mostly due to spatial index creation. That is the difference between the first and other runs. The heuristics is tricky and there are cases where it chooses slower option.

> Is the performance delta consistent across choices of inputs and operations, for both shapely and pygeos?

I doubt it is, because you can have different lengths of dfs (one aspect) different geom types (another) and different complexity of geometries (another one). Too many options.

> Will this dynamic behavior become a burden if we choose to introduce more predicates later on?

Probably. I would rather have longer documentation explaining the aspects and giving recommendations than a complex heuristics trying to guess the right settings.