@adriangb 2020-05-14
Cool, I think the path foreword is pretty clear then: we should get rid of dynamic sindex selection. 

From manual testing, the combination that results in no flipping of predicates is:
```python3
# query index
if right_df.sindex:
    l_idx, r_idx = right_df.sindex.query_bulk(
        left_df.geometry, predicate=op, sort=False
    )
    result = pd.DataFrame({"_key_left": l_idx, "_key_right": r_idx})
else:
    # when output from the join has no overlapping geometries
    # or right_df is empty / has no valid geometries
    result = pd.DataFrame(columns=["_key_left", "_key_right"], dtype=float)
```

I'll push a commit in a sec to implement this.